Reorganize db files around resources
