+++ components/db/index.ts
@@ -1,4 +1,3 @@
-export * from "./operations"
-export * from "./types"
-export * from "./hooks"
 export * from "./profile"
+export * from "./bills"
+export * from "./members"
